# refactor: replace sample_autogen code with variadic function template