This replaces the ginormous autogenerated file a few function templates that do the same (sometimes even faster). The primary goal was to make the header smaller to potentially speedup builds (as `sample.h` which includes `sample_autogen.h` is included very often) by reducing the header size. In reality, only `copyN*` (where $N = 1, 2, 3$) are used in practice so I'm also fine if we just strip everything else from the file. Apart from that the function templates (while a bit difficult to read for sure) so anything the autogenerated code did while being very slightly faster as well. 

Builds on my end (and can likely be made to build everywhere).

I would appreciate some opinion on how we want to proceed. I definitely want to throw out the functions we currently don't need and I'm open to whether we want to replace them with the variadic function template or not.