@daschuer 2024-11-25
> that do the same (sometimes even faster). 

Do you know hwere the gain for faster execution is? 

> to potentially speedup builds

Are you sure about this? From my experience templates solw down the build.