@Swiftb0y 2024-11-25
> Do you know hwere the gain for faster execution is?

Not entirely, I didn't bother digging into differences in the assembly but I did run the benchmarks to verify this didn't cause performance regressions
<details><summary>Benchmark Results</summary>
<p>


Old:
```
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
BM_Copy2WithGain/64         2388 ns         1500 ns            1
BM_Copy2WithGain/512         553 ns          435 ns            1
BM_Copy2WithGain/4096       1045 ns         1026 ns            1
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_Copy2WithRampingGain/64         2836 ns         1905 ns            1
BM_Copy2WithRampingGain/512         863 ns          720 ns            1
BM_Copy2WithRampingGain/4096       1603 ns         1527 ns            1
```

New:
```
----------------------------------------------------------------
Benchmark                      Time             CPU   Iterations
----------------------------------------------------------------
BM_Copy2WithGain/64         2451 ns         1582 ns            1
BM_Copy2WithGain/512         503 ns          459 ns            1
BM_Copy2WithGain/4096       1042 ns          978 ns            1
-----------------------------------------------------------------------
Benchmark                             Time             CPU   Iterations
-----------------------------------------------------------------------
BM_Copy2WithRampingGain/64         2630 ns         1424 ns            1
BM_Copy2WithRampingGain/512         549 ns          468 ns            1
BM_Copy2WithRampingGain/4096       1127 ns         1070 ns            1
```


</p>
</details> 

> Are you sure about this? From my experience templates solw down the build.

So I've just made some benchmark and I found that the compilation times are very similar for in the low parameter count case. It gets worse exponentally worse on GCC for more parameters, but increases linearly on clang. 
In practice, we really don't need this giant unfolded code and I would simply like to suggest all the unneeded implementations and just use a couple handwritten ones. I'm also not a huge fan of the templates either, nevertheless are they better than the autogenerated code IMO. We should only reach for autogenerated code if the compilers built-in facilities are not powerful enough (which they are in this case). In practice keeping this autogenerated code makes it infeasible to refactor which is why I really want to get rid of it.

It may be worth experimenting with removing the `if (src.gain == 0)` cascade code because thats the primary source of the template complexity and also the reason for the many instantiations on bigger inputs. Regardless, these bigger versions are not used in code anyways so its questionable how much sense it makes to argue based on the asymptotic big $N$ case.